LayoutTests/http/tests/security/seamless/seamless-sandbox-srcdoc.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test-pre.js"></script>
<div id="parent" style="width: 200px; height: 200px;">
<iframe id="iframe" seamless sandbox srcdoc="<html style='width: 100px; height: 100px; background-color: green'></html>"></iframe>
</div>
<script>
window.onload = function () {
    window.iframe = document.getElementById("iframe");
    shouldBeTrue("iframe.seamless");
    // When an iframe is sandboxed, its origin is unique, thus seamless check
    // should fail, but the spec makes a specific exception for srcdoc iframes.
    shouldBeEqualToString("window.getComputedStyle(iframe).width", "200px");
    shouldBeEqualToString("window.getComputedStyle(iframe).height", "100px");
}
</script>
