LayoutTests/http/tests/security/seamless/seamless-sandbox-srcdoc-expected.txt

FAIL iframe.seamless should be true (of type boolean). Was undefined (of type undefined).
FAIL window.getComputedStyle(iframe).width should be 200px. Was 300px.
FAIL window.getComputedStyle(iframe).height should be 100px. Was 150px.
